refactor(stdio): split into `console.rs` and `uhyve.rs`
