+++ src/fd/stdio/mod.rs
@@ -1,0 +1,5 @@
+mod console;
+mod uhyve;
+
+pub use self::console::{ConsoleStderr, ConsoleStdin, ConsoleStdout};
+pub use self::uhyve::{UhyveStderr, UhyveStdin, UhyveStdout};
